refactor: rename " Contact Card" to "Card Contact" and move him to the customs folder
