+++ src/components/customs/cards/CardContact/styles.sass
@@ -2,55 +2,58 @@
 @use "../../../../assets/sass/mixins/typography" as *
 
 // Component styling
-.contact-card
-  flex-grow: 1
+.card-contact
   display: flex
 
   &__link
     display: none
 
   &:hover &__content
-    transform: scale(1.15) translateY(-10px)
+    transform: scale(1.2) translateY(-15px)
 
   &:hover &__content &__content__label
     opacity: 1
 
   &__content
-    cursor: pointer
     position: relative
+    cursor: pointer
     display: flex
-    align-items: center
-    justify-content: center
-    width: 45px
-    height: 50px
-    background-color: $color-secondary-1
+    padding: 14px 16px
+    outline: none
     border: 0
     border-radius: 8px
-    outline: none
+    background-color: #2d334a
+    box-shadow: $shadow-primary-1
     transition: transform 500ms ease
 
     &__icon
-      font-size: 1.75rem
-      color: $color-primary-1
+      flex-grow: 1
+      display: flex
+      align-items: center
+      justify-content: center
+      font-size: 2rem
+      color: #ffffff
 
     &__label
       opacity: 0
       position: absolute
-      bottom: -12px
-      transform: translateY(100%)
+      bottom: -16px
+      left: 50%
+      transform: translate(-50%, 100%)
       padding: 4px 8px
-      @include font-custom(12, 16, center, 400, "Outfit", $color-primary-1)
-      background-color: $color-secondary-1
       border-radius: 4px
-      transition: opacity 1s ease
+      background-color: #485175
+      transition: opacity 600ms linear
+      white-space: nowrap
+      @include font-custom(12, 16, center, 400, "Outfit", #ffffff)
 
       &:before
         position: absolute
-        top: -10px
+        top: 0
         left: 50%
         content: ""
-        transform: translateX(-50%)
+        transform: translate(-50%, -90%)
         border-top: 6px solid transparent
         border-right: 6px solid transparent
-        border-bottom: 6px solid $color-secondary-1
+        border-bottom: 6px solid #485175
         border-left: 6px solid transparent
